Addressed the comments
