+++ UnitySDK/Assets/ML-Agents/Scripts/InferenceBrain/TensorApplier.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: 99d5dc2d52e442d1a1f466a246cfb28d
-timeCreated: 1539118675
+fileFormatVersion: 2
+guid: d1bef4f4ae72645108f16614355473e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
