--- UnitySDK/Assets/ML-Agents/Scripts/InferenceBrain/TensorApplierInvoker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d1bef4f4ae72645108f16614355473e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
