--- UnitySDK/Assets/ML-Agents/Scripts/InferenceBrain/TensorGeneratorInvoker.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6a24e86bc77c4a5088a5fd04d6d30e81
-timeCreated: 1537484304
